+++ CHANGELOG.md
@@ -12,6 +12,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 [@Sonicpineapple]: https://github.com/Sonicpineapple
 [@thatcomputerguy0101]: https://github.com/thatcomputerguy0101
 
+## [2.0.0-zeta.6] - 2026-02-27
+
+### Added
+
+- Recentering using middle-click, <kbd>Ctrl</kbd>+click (Windows/Linux), or <kbd>Cmd</kbd>+click (macOS)
+  - Hold <kbd>Shift</kbd> to recenter a piece instead of an axis
+  - Hold <kbd>Alt</kbd> to recenter to the back/outside instead of front
+  - <kbd>Ctrl</kbd>+right-click (Windows/Linux) or <kbd>Cmd</kbd>+right-click (macOS) to do the reverse rotation
+
+### Fixed
+
+- Crash when window starts too small
+
 ## [2.0.0-zeta.5] - 2026-02-27
 
 ### Fixed